Completed JS portion

+++ src/components/buttons/DeleteButton.js
@@ -1,44 +1,52 @@
 import React from 'react';
 import { withRouter } from 'react-router-dom';
+import { connect } from 'react-redux';
+import * as actions from '../../actions';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrashAlt } from '@fortawesome/free-solid-svg-icons';
 import { storageRef, database } from '../../services/firebaseConfig';
 
 class DeleteButton extends React.Component {
-  deleteItem = (trailId) => {
-    console.log('delete button working for: ' + trailId);
+  deleteItem = (trailId, trailPhotoId, itemStatus) => {
+    this.props.updateTrailItem(itemStatus);
 
-    const itemStatus = this.props.itemStatus;
-
-    console.log('delete button category: ' + itemStatus);
-
-    const photoRef = storageRef.child(`trailphotos/${trailId}`);
+    if (trailPhotoId) {
+      const photoRef = storageRef.child(`trailphotos/${trailId}`);
+      photoRef
+        .delete()
+        .then(function () {
+          console.log('Photo removal succeeded.');
+        })
+        .catch(function (error) {
+          alert('Error removing photo!');
+          console.log('Error removing photo!');
+        });
+    }
 
     database
       .ref(`/${itemStatus}items/` + trailId)
       .remove()
       .then(function () {
         console.log('Remove active succeeded.');
-      });
-
-    photoRef
-      .delete()
-      .then(function () {
-        console.log('Photo removal succeeded.');
+      })
+      .then(() => {
+        alert('Item Deleted!');
       })
       .catch(function (error) {
-        console.log('Error removing photo!');
+        alert('Error removing item!');
       });
-    this.props.history.push(`trailworklist/${itemStatus}items`);
+    this.props.history.push(`/tablelist/${itemStatus}items`);
   };
 
   render() {
+    const { trailId, trailPhotoId, itemStatus } = this.props;
+
     return (
       <button
         type='button'
         className='btn'
         aria-label='delete'
-        onClick={() => this.deleteItem(this.props.trailId)}
+        onClick={() => this.deleteItem(trailId, trailPhotoId, itemStatus)}
       >
         <FontAwesomeIcon icon={faTrashAlt} />
       </button>
@@ -46,4 +54,10 @@ class DeleteButton extends React.Component {
   }
 }
 
-export default withRouter(DeleteButton);
+const mapStateToProps = (state) => {
+  return {
+    trailItems: state.trailData.trailItems,
+  };
+};
+
+export default connect(mapStateToProps, actions)(withRouter(DeleteButton));
